Add 3 version of SHM
First -> primary version. This version is for advanced course. Show few
design patterns and how to write simple and easy to expand code.

Second -> advnced version -> This is modified primary version. Deleted
some design pattern and create hard to read and maintenance code.

Third -> For beginer -> This code has example form lesson for the
beginers. This code shouldn't be pretty. The basic idea is to show how
objecting progrming wokr.